added parent class, deleted .* from import
