+++ src/test/java/com/epam/tc/hw2/ExerciseOneTest.java
@@ -1,56 +1,21 @@
 package com.epam.tc.hw2;
 
-import io.github.bonigarcia.wdm.WebDriverManager;
 import java.time.Duration;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 import org.assertj.core.api.Assertions;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
-import org.testng.asserts.SoftAssert;
 
-public class ExerciseOneTest {
+public class ExerciseOneTest extends AbstractTest {
 
-    static String XPATH_PASSWORD_INPUT = "//input[@id='password']";
-    static String XPATH_LOGIN_SUBMIT = "//button[@id='login-button']";
-    static String XPATH_LOGOUT_SUBMIT = "//*[@class='logout']";
-    static String ID_LOGIN_NAME = "user-name";
-    static String ID_FRAME = "frame";
-    static String loginText = "Roman";
-    static String passwordText = "Jdi1234";
-    static String HOME_PAGE = "Home Page";
-    static String LOGIN_ROMAN_IOVLEV = "ROMAN IOVLEV";
-    static String XPATH_LOGIN_DROPDOWN = "//*[@class='dropdown uui-profile-menu']";
-    static String XPATH_LOGIN_INPUT = "//input[@id='name']";
-    public static WebDriver driver;
-    private static WebDriverWait webDriverWait;
-    private static SoftAssert softAssert;
+
     private static String XPATH_ITEMS_IN_HEADER = "//ul[@class=\"uui-navigation nav navbar-nav m-l8\"]";
     private static String XPATH_LEFT_MENU = "//*[@class='sidebar-menu left']";
     private static String ID_FRAME_BUTTON = "frame-button";
     private static String LIST_ITEMS_HEADER = "HOME\nCONTACT FORM\nSERVICE\nMETALS & COLORS";
-    public static String URL_HOME_PAGE = "https://jdi-testing.github.io/jdi-light/index.html";
-
-    @BeforeTest
-    static void setupAll() {
-        WebDriverManager.chromedriver().setup();
-    }
-
-    @BeforeMethod(alwaysRun = true)
-    public static void browserDriverSetup() {
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-        driver.manage().timeouts().implicitlyWait(2, TimeUnit.SECONDS);
-        softAssert = new SoftAssert();
-    }
 
     @Test
     public void exercise() {
@@ -117,11 +82,5 @@ public class ExerciseOneTest {
                 .contains("Service")
                 .contains("Metals & Colors")
                 .contains("Elements packs");
-    }
-
-    @AfterMethod(alwaysRun = true)
-    public void browserQuit() {
-        driver.quit();
-        driver = null;
     }
 }
